Break Reporter into CcsReporter and EigenReporter
