+++ elk/training/eigen_reporter.py
@@ -1,31 +1,16 @@
 """An ELK reporter network."""
 
 from ..math_util import batch_cov
+from .reporter import Reporter, ReporterConfig
 from dataclasses import dataclass
-from pathlib import Path
-from simple_parsing.helpers import Serializable
-from sklearn.metrics import roc_auc_score
 from torch import Tensor
-from typing import Literal, NamedTuple, Optional, Union
+from typing import Literal, Optional
 import torch
 import torch.nn as nn
 
 
-class EvalResult(NamedTuple):
-    """The result of evaluating a reporter on a dataset.
-
-    The `.score()` function of a reporter returns an instance of this class,
-    which contains the loss, accuracy, calibrated accuracy, and AUROC.
-    """
-
-    loss: float
-    acc: float
-    cal_acc: float
-    auroc: float
-
-
 @dataclass
-class EigenReporterConfig(Serializable):
+class EigenReporterConfig(ReporterConfig):
     """ """
 
     inv_weight: float = 1.0
@@ -33,13 +18,15 @@ class EigenReporterConfig(Serializable):
     solver: Literal["arpack", "dense", "power"] = "dense"
 
 
-class EigenReporter(nn.Module):
+class EigenReporter(Reporter):
     """A linear reporter whose weights are computed via eigendecomposition.
 
     Args:
         in_features: The number of input features.
         cfg: The reporter configuration.
     """
+
+    config: EigenReporterConfig
 
     def __init__(
         self, in_features: int, cfg: EigenReporterConfig, device: Optional[str] = None
@@ -49,51 +36,28 @@ class EigenReporter(nn.Module):
         self.config = cfg
         self.linear = nn.Linear(in_features, 1, bias=False, device=device)
 
-    # TODO: These methods will do something fancier in the future
-    @classmethod
-    def load(cls, path: Union[Path, str]):
-        """Load a reporter from a file."""
-        return torch.load(path)
-
-    def save(self, path: Union[Path, str]):
-        # TODO: Save separate JSON and PT files for the reporter.
-        torch.save(self, path)
-
     def forward(self, x: Tensor) -> Tensor:
         """Return the raw score output of the probe on `x`."""
-        return self.linear(x)
+        return self.linear(x).squeeze(-1)
 
-    def validate_data(self, data):
-        """Validate that the data's shape is valid."""
-        assert len(data) == 2 and data[0].shape == data[1].shape
+    def predict(self, x_pos: Tensor, x_neg: Tensor) -> Tensor:
+        return 0.5 * (self(x_pos) - self(x_neg))
 
     def fit(
         self,
-        contrast_pair: tuple[Tensor, Tensor],
+        x_pos: Tensor,
+        x_neg: Tensor,
+        labels: Optional[Tensor] = None,
     ) -> float:
-        """Fit the probe to the contrast pair (x0, x1).
+        """Fit the probe to the contrast pair (x_pos, x_neg).
 
         Args:
-            contrast_pair: A tuple of tensors, (x0, x1), where x0 and x1 are the
-                contrastive representations.
-            labels: The labels of the contrast pair. Defaults to None.
-            lr: The learning rate for Adam. Defaults to 1e-2.
-            num_epochs: The number of epochs to train for. Defaults to 1000.
-            num_tries: The number of times to repeat the procedure. Defaults to 10.
-            optimizer: The optimizer to use. Defaults to "adam".
-            verbose: Whether to print out information at each step. Defaults to False.
-            weight_decay: The weight decay for Adam. Defaults to 0.01.
+            x_pos: The positive examples.
+            x_neg: The negative examples.
 
         Returns:
-            best_loss: The best loss obtained.
-
-        Raises:
-            ValueError: If `optimizer` is not "adam" or "lbfgs".
-            RuntimeError: If the best loss is not finite.
+            loss: The best loss obtained.
         """
-        self.validate_data(contrast_pair)
-
-        x_pos, x_neg = contrast_pair
         assert x_pos.shape == x_neg.shape
 
         # Variance
@@ -113,48 +77,3 @@ class EigenReporter(nn.Module):
         self.linear.weight.data = Q[:, -1, None]
 
         return L[-1]
-
-    @torch.no_grad()
-    def score(
-        self,
-        contrast_pair: tuple[Tensor, Tensor],
-        labels: Tensor,
-    ) -> EvalResult:
-        """Score the probe on the contrast pair (x0, x1).
-
-        Args:
-            contrast_pair: A tuple of tensors, (x0, x1), where x0 and x1 are the
-                contrastive representations.
-            labels: The labels of the contrast pair.
-
-        Returns:
-            an instance of EvalResult containing the loss, accuracy, calibrated
-                accuracy, and AUROC of the probe on the contrast pair (x0, x1).
-        """
-
-        self.validate_data(contrast_pair)
-
-        logit0, logit1 = map(self, contrast_pair)
-        p0, p1 = logit0.sigmoid(), logit1.sigmoid()
-        pred_probs = 0.5 * (p0 + (1 - p1))
-
-        # Calibrated accuracy
-        cal_thresh = pred_probs.float().quantile(labels.float().mean())
-        cal_preds = pred_probs.gt(cal_thresh).squeeze(1).to(torch.int)
-        raw_preds = pred_probs.gt(0.5).squeeze(1).to(torch.int)
-
-        # makes `num_variants` copies of each label, all within a single
-        # dimension of size `num_variants * n`, such that the labels align
-        # with pred_probs.flatten()
-        broadcast_labels = labels.repeat_interleave(pred_probs.shape[1])
-        # roc_auc_score only takes flattened input
-        auroc = float(roc_auc_score(broadcast_labels.cpu(), pred_probs.cpu().flatten()))
-        cal_acc = cal_preds.flatten().eq(broadcast_labels).float().mean()
-        raw_acc = raw_preds.flatten().eq(broadcast_labels).float().mean()
-
-        return EvalResult(
-            loss=0.0,
-            acc=torch.max(raw_acc, 1 - raw_acc).item(),
-            cal_acc=torch.max(cal_acc, 1 - cal_acc).item(),
-            auroc=max(auroc, 1 - auroc),
-        )
